.claude/plans/merge-convert-skills/tools/ir-validate/pyproject.toml
[project]
name = "ir-validate"
version = "1.0.0"
description = "Validation tool for IR (Intermediate Representation) documents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ir", "validation", "code-conversion", "schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
full = [
    "jsonschema>=4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.8",
]

[project.scripts]
ir-validate = "ir_validate.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["ir_validate"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "jsonschema",
    "yaml",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
